Uses Big decimal and Bankers rounding to show amounts in app. TODO fix Moshi adapter for BigDecimal. Adds more documentation

+++ app/src/main/java/com/gnb/gnbtrades/domain/entities/Trade.kt
@@ -1,5 +1,6 @@
 package com.gnb.gnbtrades.domain.entities
 
+import java.math.BigDecimal
 import java.util.*
 
-data class Trade (val amount: Float, val currency : String)
+data class Trade (val amount: BigDecimal, val currency : String)
